feat: move entity e2e tests to dedicated file
