+++ tests/e2e/entities.spec.js
@@ -1,0 +1,15 @@
+import { test, expect } from '@playwright/test';
+
+test('add entities to the canvas and change name', async ({ page }) => {
+    await page.goto('/');
+
+    const canvas = page.locator("svg")
+    await page.getByRole('img').first().dragTo(canvas);
+
+    await canvas.click();
+
+    await page.getByText('Entidad').first().dblclick();
+    await page.keyboard.type('Clientes');
+
+    await canvas.click();
+});
